contract: logic_admin merge into storage
